working with some separation of files for utils
